+++ src/lib/pages/public-project/components/tables/contracts/PublicProjectContractRow.tsx
@@ -13,6 +13,7 @@ import {
   useInternalNavigate,
 } from "lib/app-provider";
 import { AppLink } from "lib/components/AppLink";
+import { ContractInteractionTabs } from "lib/components/ContractInteractionSwitch";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { CustomIcon } from "lib/components/icon";
 import {
@@ -97,14 +98,14 @@ export const PublicProjectContractRow = ({
           onClick={(e) => e.stopPropagation()}
         >
           <AppLink
-            href={`/execute?contract=${publicContractInfo.publicInfo.contractAddress}`}
+            href={`/contract-interaction?selectedType=${ContractInteractionTabs.EXECUTE}contract=${publicContractInfo.publicInfo.contractAddress}`}
           >
             <Button variant="outline-gray" size="sm">
               Execute
             </Button>
           </AppLink>
           <AppLink
-            href={`/query?contract=${publicContractInfo.publicInfo.contractAddress}`}
+            href={`/contract-interaction?selectedType=${ContractInteractionTabs.QUERY}&contract=${publicContractInfo.publicInfo.contractAddress}`}
           >
             <Button variant="outline-gray" size="sm">
               Query